# services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta API permite interações com dispositivos e usuários do sistema AgroTech",
        "title": "API para Sistema AgroTech",
        "contact": {
            "name": "arthurcadore",
            "url": "arthurcadore.github.io",
            "email": "arthurbarcella.ifsc@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/cooperativas": {
            "get": {
                "description": "Obtém a lista de todas as cooperativas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cooperativas"
                ],
                "summary": "Lista cooperativas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Cooperativa"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma nova cooperativa no banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cooperativas"
                ],
                "summary": "Cria uma nova cooperativa",
                "parameters": [
                    {
                        "description": "Nova Cooperativa",
                        "name": "cooperativa",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Cooperativa"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Cooperativa"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cooperativas/{cnpj}": {
            "delete": {
                "description": "Deleta uma cooperativa com base no CNPJ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cooperativas"
                ],
                "summary": "Deleta uma cooperativa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CNPJ da Cooperativa",
                        "name": "cnpj",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deletado com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/devices": {
            "get": {
                "description": "Obtém a lista de todos os dispositivos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Lista dispositivos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Device"
                            }
                        }
                    }
                }
            }
        },
        "/devices/{id}": {
            "delete": {
                "description": "Deleta um dispositivo do banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Deleta um dispositivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do dispositivo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/devices/{uuid}": {
            "get": {
                "description": "Obtém os dados de um dispositivo específico pelo UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Lista dados do dispositivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do dispositivo",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Data"
                            }
                        }
                    }
                }
            }
        },
        "/devices/{uuid}/sensores/{id}": {
            "delete": {
                "description": "Deleta um sensor de um dispositivo com base no UUID do dispositivo e ID do sensor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sensores"
                ],
                "summary": "Deleta um sensor de um dispositivo específico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do dispositivo",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID do sensor",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/map": {
            "get": {
                "description": "Retorna os dispositivos com as colunas UUID, LATITUDE, LONGITUDE e PESO",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Lista todos os dispositivos com localização e peso",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.DeviceMapResponse"
                            }
                        }
                    }
                }
            }
        },
        "/sensores": {
            "get": {
                "description": "Obtém a lista de todos os sensores",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sensores"
                ],
                "summary": "Lista sensores",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Sensors"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo sensor no banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sensores"
                ],
                "summary": "Cria um novo sensor",
                "parameters": [
                    {
                        "description": "Novo Sensor",
                        "name": "sensor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Sensors"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Sensors"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Obtém a lista de usuários",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Lista usuários",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo usuário no banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Cria um novo usuário",
                "parameters": [
                    {
                        "description": "Novo Usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/{matricula}": {
            "delete": {
                "description": "Deleta um usuário do banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Deleta um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matricula do usuário",
                        "name": "matricula",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.DeviceMapResponse": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "peso": {
                    "type": "number"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "models.Cooperativa": {
            "type": "object",
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "endereco": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "models.Data": {
            "type": "object",
            "properties": {
                "id_sensor": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "valor": {
                    "type": "number"
                }
            }
        },
        "models.Device": {
            "type": "object",
            "properties": {
                "altitude": {
                    "type": "number"
                },
                "firmware_version": {
                    "type": "string"
                },
                "hardware_version": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "peso": {
                    "type": "number"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "models.Sensors": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "tipo": {
                    "type": "string"
                },
                "unidade": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "licencas": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        }
    }
}